Moved QRCodeGenerator to sample app from library
