+++ iroha-android-sample/src/main/java/io/soramitsu/iroha/util/QRCodeGenerator.java
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package io.soramitsu.irohaandroid.qr;
+package io.soramitsu.iroha.util;
 
 import android.graphics.Bitmap;
 import android.graphics.Color;
